requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.15.0
Pillow==10.0.0
opencv-python==4.8.1.78
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
streamlit-chat==0.1.0
altair==5.0.1
python-dotenv==1.0.0
faker==19.0.0